API change: Overhaul priority. Add Requeue()
Adding Requeue() method to insert a message back into the queue with its
precedence intact. This has the same behavior that Enqueue() had when
dealing with messages that had come out of the queue, only, now that
priority is decoupled from the message itself (see below). This means
that  messages can be inserted into the queue with Requeue() at a
different priority than they exited the queue with. Since the message
re-uses its key, it will exit the queue  before newer messages with
the same priority.

Correspondingly, Enqueue() will ALWAYS treat messages as new by
assigning a new key, so inserting a message with Enqueue() will always
place it behind other messages with that priority,

The message's priority is now no longer an intrinsic property in
enqueueing. Rather than being created with the message, Priority is
specified when enqueueing or requeueing (see below). The Priority()
method has been added to messages so that the user can see what priority
the message emerged from the queue with. It will be -1 if the message is
new and has not been enqueued.